Allocation test upgrades
