+++ Tests/Editor/RegistrationStorageStructuralGuardTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31f5d3e82bcd4ed5bf578b07c826345a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
